Shift JSONTransformer and XMLTransformer to new modules
